+++ config.ini
@@ -1,0 +1,3 @@
+[WINDOW]
+width = 1080
+height = 640
